audioflinger: put fastmixer with active tracks into mix state when suspended
Fastmixer will enter COLD_IDLE state after the output is suspended.
If it still has active tracks and may be removed later, the state
command sent to fastmixer could be blocked and thus block subsequent
calls to this output.
Put fastmixer into MIX state instead of COLD_IDLE when it's suspended
but still has active tracks to be processed.

Change-Id: I7f24f1a3e5b1651001fcef3ba3d00a8fb67e2b79
CRs-Fixed: 1093067